requirements.txt
# Python dependencies for Instagram bot
selenium==4.15.2
python-dotenv==1.0.0
google-generativeai==0.3.1
requests==2.31.0
user-agents==2.2.0
pillow==10.1.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
loguru==0.7.2